changed type of Ean from int to long to match the OrgUnit typ

+++ Core.ApplicationServices/Model/Organizations/Write/OrganizationUnitUpdateParameters.cs
@@ -10,7 +10,7 @@ namespace Core.ApplicationServices.Model.Organizations.Write
         public OptionalValueChange<string> Name { get; set; }
         public OptionalValueChange<OrganizationUnitOrigin> Origin { get; set; }
         public OptionalValueChange<Maybe<Guid>> ParentUuid { get; set; }
-        public OptionalValueChange<Maybe<int>> Ean { get; set; }
+        public OptionalValueChange<Maybe<long>> Ean { get; set; }
         public OptionalValueChange<Maybe<string>> Id { get; set; }
     }
 }
